Rewrite stopPropagation to TS
